feat: move trie and util to internal and fix adapter type
